Add python sample code + tests
Code stolen from my giftig/s3-browser project

Tweak the import regex a bit as it was failing to find single python
imports: that's because it was expecting a space directly after import,
and then either a dotted path, import block, or additional space
directly before the symbol. In a simple case that'd require two spaces,
which isn't correct. Fix this by simply removing the space after the
word import; it's still going to be looking for a block or dotted path
next anyway.

This adjustment technically means it might incorrectly match a string
like "import.foo.bar.baz", but that seems unlikely to come up.