Split out SearchStrategy
A search strategy will take a series of patterns to match on, so that in
future we'll be able to apply different strategies to different
languages.

Also define a SmartMode as golang needs a separate mode, and this
provides better separation than just matching on language within the
Search. Golang will therefore define its own SearchStrategy even though
currently it's only deifnining its own SmartMode definition, not
changing the patterns it's using.

get_strategy will take language and return the go strategy for go, and
the default strategy for everything else. This can be further separated
later if needed.